gcc 4.4.1: fixing some compiler warnings
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@1023 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ test_suite/test_spat_object_trees.cpp
@@ -346,7 +346,7 @@ void testSpatObjectTrees()
 	{
 		tree.remove(i);
 	}
-	LASS_TEST_CHECK_EQUAL(tree.objectCount(), 0);
+	LASS_TEST_CHECK_EQUAL(tree.objectCount(), size_t(0));
 }
